fix(#1020): Codex sweep BLOCKING — _current refresh + savepoints + BIGINT
Second-pass fixes over PRs #1037-#1040 closing Codex sweep BLOCKING items:

**C3/C4/C5 _current refresh wiring** — bulk ingesters write to
ownership_*_observations only; rollup endpoints read from _current.
Without this wiring, AAPL/MSFT show 0 ownership after a successful
bulk ingest. Each ingester now accumulates touched_instrument_ids;
the orchestrator job loops them and calls refresh_institutions_current
/ refresh_insiders_current / refresh_funds_current. Refresh failures
RAISE before disk cleanup so archives stay available for retry.

**Insider + N-PORT per-row savepoint** — record_insider_observation
and record_fund_observation were called outside any savepoint. One
CHECK violation flipped psycopg into InFailedSqlTransaction and every
subsequent row in the file silently failed until commit. Both ingesters
now wrap each write in `with conn.transaction()` matching the C3 13F
pattern.

**INTEGER → BIGINT widening** (sql/133) — ownership_*_observations and
_current tables had `instrument_id INTEGER`, but the parent
`instruments.instrument_id` is BIGINT. Widening prevents future
truncation and lets the planner skip implicit casts on join columns.
Source CREATE TABLE definitions also updated for fresh installs.

**Per-archive transient retry** (#1031 follow-up) — `_download_one_with_retry`
wraps each archive download with 3 attempts + exponential backoff on
ConnectError / ReadTimeout / RemoteProtocolError. Closes the silent
"2/18 archives missing" failure mode observed on the first live run.

**Per-archive deferred deletion** — C3/C4/C5 used to delete each
archive immediately on per-archive success. Refactored to defer
deletion until END of stage success so a later archive failure does
not strand prior archives' retry input.

**CUSIP map preload** — replaced per-row `_resolve_cusip(conn, cusip)`
DB query in 13F + N-PORT ingesters with `_load_cusip_map(conn)` dict
preload. Multi-million-row INFOTABLE.tsv was DB-round-trip bound.

**Stage requires** — legacy chain stages (filings_history_seed,
sec_first_install_drain, sec_insider_transactions_backfill, sec_form3_ingest,
sec_13f_recent_sweep, sec_n_port_ingest) now depend on cik_refresh
(was universe_sync); their per-CIK fetches need the CIK mapping.

Test updates:
- test_sec_bulk_download.TestPreflightCleanup: assertion updated to
  match new run-manifest contract (preflight wipes BOTH .partial AND
  complete .zip — see _preflight_cleanup_stale_partials docstring).
- test_bootstrap_flow_integration.test_bootstrap_partial_error_then_retry_failed:
  assertion updated to match same-lane-only retry semantics from the
  Phase C concurrent dispatch refactor (#1039).

Follow-ups filed: #1041 (slow-connection fallback manifest), #1042 (A3
shared SEC rate clock), #1043 (CHECK constraints), #1044
(first_install_drain refactor), #1045 (PipelinedSecFetcher D-stages),
#1046 (frontend archive-level progress).

Closes #1020 sweep findings (BLOCKING-only).

+++ sql/114_ownership_institutions_observations.sql
@@ -31,7 +31,7 @@
 BEGIN;
 
 CREATE TABLE IF NOT EXISTS ownership_institutions_observations (
-    instrument_id           INTEGER NOT NULL,
+    instrument_id           BIGINT NOT NULL,
     filer_cik               TEXT NOT NULL,
     filer_name              TEXT NOT NULL,
     filer_type              TEXT
@@ -104,7 +104,7 @@ CREATE INDEX IF NOT EXISTS idx_inst_obs_filer_period
 
 
 CREATE TABLE IF NOT EXISTS ownership_institutions_current (
-    instrument_id           INTEGER NOT NULL,
+    instrument_id           BIGINT NOT NULL,
     filer_cik               TEXT NOT NULL,
     filer_name              TEXT NOT NULL,
     filer_type              TEXT
